+++ web/tests/dispatch.test.cjs
@@ -539,9 +539,6 @@ run_test("realm settings", ({override}) => {
     event = event_fixtures.realm__update__invite_required;
     test_realm_boolean(event, "realm_invite_required");
 
-    event = event_fixtures.realm__update__invite_to_realm_policy;
-    test_realm_integer(event, "realm_invite_to_realm_policy");
-
     event = event_fixtures.realm__update__want_advertise_in_communities_directory;
     test_realm_boolean(event, "realm_want_advertise_in_communities_directory");
 
@@ -605,6 +602,7 @@ run_test("realm settings", ({override}) => {
     override(realm, "realm_authentication_methods", {Google: {enabled: false, available: true}});
     override(realm, "realm_can_add_custom_emoji_group", 1);
     override(realm, "realm_can_create_public_channel_group", 1);
+    override(realm, "realm_can_invite_users_group", 1);
     override(realm, "realm_can_move_messages_between_topics_group", 1);
     override(realm, "realm_direct_message_permission_group", 1);
     override(realm, "realm_plan_type", 2);
@@ -620,6 +618,7 @@ run_test("realm settings", ({override}) => {
     });
     assert_same(realm.realm_can_add_custom_emoji_group, 3);
     assert_same(realm.realm_can_create_public_channel_group, 3);
+    assert_same(realm.realm_can_invite_users_group, 3);
     assert_same(realm.realm_can_move_messages_between_topics_group, 3);
     assert_same(realm.realm_direct_message_permission_group, 3);
     assert_same(realm.realm_plan_type, 3);