+++ web/tests/lib/events.cjs
@@ -302,13 +302,6 @@ exports.fixtures = {
         value: false,
     },
 
-    realm__update__invite_to_realm_policy: {
-        type: "realm",
-        op: "update",
-        property: "invite_to_realm_policy",
-        value: 2,
-    },
-
     realm__update__invite_to_stream_policy: {
         type: "realm",
         op: "update",
@@ -371,6 +364,7 @@ exports.fixtures = {
             },
             can_add_custom_emoji_group: 3,
             can_create_public_channel_group: 3,
+            can_invite_users_group: 3,
             can_move_messages_between_topics_group: 3,
             direct_message_permission_group: 3,
             plan_type: 3,