+++ zerver/lib/events.py
@@ -589,7 +589,7 @@ def fetch_initial_state_data(
             or state["can_create_web_public_streams"]
         )
         state["can_subscribe_other_users"] = settings_user.can_subscribe_other_users()
-        state["can_invite_others_to_realm"] = settings_user.can_invite_users_by_email()
+        state["can_invite_others_to_realm"] = settings_user.can_invite_users_by_email(realm)
         state["is_admin"] = settings_user.is_realm_admin
         state["is_owner"] = settings_user.is_realm_owner
         state["is_moderator"] = settings_user.is_moderator
@@ -1302,26 +1302,11 @@ def apply_event(
                     else state["server_jitsi_server_url"]
                 )
 
-            policy_permission_dict = {
-                "invite_to_stream_policy": "can_subscribe_other_users",
-                "invite_to_realm_policy": "can_invite_others_to_realm",
-            }
-
-            # Tricky interaction: Whether we can create streams and can subscribe other users
-            # can get changed here.
-
-            if field == "realm_waiting_period_threshold":
-                for policy, permission in policy_permission_dict.items():
-                    if permission in state:
-                        state[permission] = user_profile.has_permission(policy)
-
             if (
-                event["property"] in policy_permission_dict
-                and policy_permission_dict[event["property"]] in state
+                event["property"] == "invite_to_stream_policy"
+                and "can_subscribe_other_users" in state
             ):
-                state[policy_permission_dict[event["property"]]] = user_profile.has_permission(
-                    event["property"]
-                )
+                state["can_subscribe_other_users"] = user_profile.has_permission(event["property"])
         elif event["op"] == "update_dict":
             for key, value in event["data"].items():
                 if key == "max_file_upload_size_mib":
@@ -1375,6 +1360,11 @@ def apply_event(
                         state["can_create_private_streams"]
                         or state["can_create_public_streams"]
                         or state["can_create_web_public_streams"]
+                    )
+
+                if key == "can_invite_users_group" and "can_invite_others_to_realm" in state:
+                    state["can_invite_others_to_realm"] = user_profile.has_permission(
+                        "can_invite_users_group"
                     )
 
                 if key == "plan_type":